test_qr_code/requirements.txt
Pillow==10.0.1
python-barcode==0.15.1
